virtiofs: Support for remote Inotify/Fsnotify events
Adding support for remote inotify/fsnotify remote events for virtiofs.
There are two modes supported one with only local events and one with
only remote events. The two modes are enabled depending on the features
that the guest client has negotiated with the fuse server.

The only remote inotify/fsnotify events affect only the exported
directory of the host through fuse and not the local filesystem of the
guest.

Signed-off-by: Ioannis Angelakopoulos <iangelak@redhat.com>