virtiofs: Add the fuse_fsnotify_send_request to FUSE
The function "fuse_fsnotify_send_request" is responsible for sending an
fsnotify FUSE request to the FUSE server.

The request contains all the information that are stored within the a
"fuse_notify_fsnotify_in" struct which is the event mask for the inotify
watch, the action to be performed on the watch (create, modify or delete)
and the group identifier which is essentially the 64bit (fsnotify_group)
pointer that corresponds to one inotify instance created by a user space process.

Since each group identifier is unique, the FUSE server can create an equal
number of inotify instances to the ones created by the user space
processes.

Signed-off-by: Ioannis Angelakopoulos <iangelak@redhat.com>