virtiofs: Add the fuse_fsnotify_update_mark inode operation
Add a new inode operation "fuse_fsnotify_update_mark", which is specific to
FUSE inodes. This operation is called when a process tries to add, modify
or remove a watch from a FUSE inode, when the remote fsnotify support is
enabled both in the kernel and the FUSE server.

Specifically, this function will call the "fuse_fsnotify_send_request"
function, to send an fsnotify request to the FUSE server related to an
inode watch.

Signed-off-by: Ioannis Angelakopoulos <iangelak@redhat.com>