virtiofs: Patch inotify and fsnotify with the remote fsnotify support
Modify the "inotify_add_watch" system call and add the support for the
remote fsnotify add/modify watch operations for FUSE inodes.

Essentially, when the kernel adds/modifies a watch locally, also send a
fsnotify request to the FUSE server to do the same. We keep the local watch
placement since it is essential for the functionality of the fsnotify
notification subsystem. However, the local events generated by the kernel
will be suppressed if they affect FUSE inodes and the remote fsnotify
support is enabled.

Also modify the "fsnotify_detach_mark" function in fs/notify/mark.c to add
support for the remote deletion of watches for FUSE inodes. In contrast to
the add/modify operation we do not modify the inotify subsystem, but the
fsnotify subsystem. That is because there are two ways of deleting a watch
from an inode. The first is by manually calling the "inotify_rm_watch"
system call and the second is automatically when the process that has
created an inotify instance exits. In that case the kernel is responsible
to delete all the watches corresponding to said inotify instance.

Thus we send the fsnotify request for the deletion of the remote watch at
the lowest level within "fsnotify_detach_mark" to catch both watch removal
cases.

Signed-off-by: Ioannis Angelakopoulos <iangelak@redhat.com>