virtiofs: Add the fuse_fsnotify_remote inode operation
Add a new inode operation "fuse_fsnotify_remote" to check if the remote
fsnotify support is enabled.

Also add a new watch bit "FS_FSNOTIFY_REMOTE" to the fsnotify watch bits.
When this bit is set on a watch mask then it means that the event orginated
from the FUSE server and not the kernel.

Signed-off-by: Ioannis Angelakopoulos <iangelak@redhat.com>