FUSE: Add the fsnotify event bits to FUSE
To introduce the support of fsnotify/inotify events to FUSE we add the
fsnotify event bits to the fuse.h header file. Similar to fsnotify these
bits will be used to create/modify/delete watches on FUSE inodes.

We redefine and integrate the fsnotify event bits into the FUSE protocol
so that the FUSE server and the client have a common concept about the
events transmitted between the two.

Signed-off-by: Ioannis Angelakopoulos <iangelak@redhat.com>