vfs,FUSE: Add the fuse_fsnotify_update_mark inode operation
Add a new inode operation "fuse_fsnotify_update_mark", which is specific to
FUSE inodes. This operation is called from the "inotify_add_watch"
system call in the inotify subsystem.

Specifically, the operation is called when a process tries to add, modify
or remove a watch from a FUSE inode and the remote fsnotify support is
enabled both in the guest kernel and the FUSE server (virtiofsd).

The "fuse_fsnotify_update_mark" function in turn calls the
"fuse_fsnotify_send_request" function, to send an fsnotify request to the
FUSE server related to an inode watch.

Signed-off-by: Ioannis Angelakopoulos <iangelak@redhat.com>