FUSE,Fsnotify: Add the fuse_fsnotify_remote inode operation
Add a new inode operation "fuse_fsnotify_remote", called by the "fsnotify"
function in fsnotify.c to check if the remote fsnotify/inotify support is
enabled and if the event generated for the target inode is a local or
remote event.

Also add a new watch bit "FS_FSNOTIFY_REMOTE" to the fsnotify watch bits.
When this bit is set on a watch mask then it means that the event orginated
from the FUSE server and not the guest kernel. The check takes place in
the "fuse_fsnotify_remote" function.

Signed-off-by: Ioannis Angelakopoulos <iangelak@redhat.com>