Fsnotify: Add a wrapper around the fsnotify function
Add a wrapper around the "fsnotify" function in fs/notify/fsnotify.c that
first checks if the event that is about to be generated is related to a
FUSE inode or not.

If yes then it checks if the remote fsnotify/inotify support is enabled
and if that is the case it calls the "fsnotify_remote" function for the
inode to check if the event should be suppressed (dropped) or not. The
event is suppressed in the case the mask does not have the
FS_FSNOTIFY_REMOTE bit set, meaning that it was generated locally by
the kernel.

The wrapper will disable the FS_FSNOTIFY_REMOTE bit from the mask after the
check is complete so that the whole process remains transparent to user
space.

Signed-off-by: Ioannis Angelakopoulos <iangelak@redhat.com>