virtiofs: Add support for handling the remote fsnotify notifications
FUSE and specifically virtiofs should be able to handle the asynchronous
event notifications originating from the FUSE server. To this end we add
the FUSE_NOTIFY_FSNOTIFY switch case to the "virtio_fs_handle_notify" in
fs/fuse/virtio_fs.c to handle these specific notifications.

The event notification contains the information that a user space
application would receive when monitoring an inode for events. The
information is the mask of the inode watch, a file name corresponding to
the inode the remote event was generated for and the inotify cookie.

Then a new event should be generated corresponding to the event
notification received from the FUSE server. To differentiate between these
custom generated "remote" events and the local events generated by the
kernel we enable temporarily the FS_NOTIFY_REMOTE bit in the inode watch
mask received from the FUSE server which will be disabled by the
"fuse_fsnotify_remote" function later.

Signed-off-by: Ioannis Angelakopoulos <iangelak@redhat.com>