Use of generation and simple Fanotify test
Add the generation field to the data transmitted between the guest and the
server. The generation is functioning as an additional measure of ID of
the target inode.

Hardcode the fsid to FUSE so that Fanotify with file handles can be used
locally by the guest. Currently the prototype will not work for all the
fanotify event, just the basic ones like FAN_OPEN or FAN_CLOSE.

The fsnotify implementation needs more work to support both inotify and
fanotify at the same time.