btrfs: Annotate the reservation space wait event with lockdep
Add a lockdep annotation for the reservation space wait event.

This wait event is special in the sense that its condition modification and
waitqueue signaling are protected by the space_info->lock spinlock.

However, that is not the case with flush_space() in fs/btrfs/space-info.c
which is usually part of the execution context that leads to signaling of
the reservation space waitqueue. This function is not protected by the
space_info->lock thus we place our annotation in flush_space().

Since, flush_space() operates differently depending on the flushing state,
we use a multilevel lockdep map where each subclass/level corresponds to
each flushing state.

Signed-off-by: Ioannis Angelakopoulos <iangelak@fb.com>