btrfs: Macros for multilevel annotations of wait events with lockdep
The new macros are similar to the generic lockdep annotation macros used to
annotate wait events in btrfs in fs/btrfs/ctree.h. However, the new macros
take as a parameter the subclass of the lockdep map, which can be used for
a multilevel annotation of a wait event.

The new macros introduced are:
  1) btrfs_might_wait_for_event_nested()
  2) btrfs_lockdep_acquire_nested()

Signed-off-by: Ioannis Angelakopoulos <iangelak@fb.com>